Extract helpers for site_channels from general channels_helpers
